+++ src/pages/Guideline/Guideline.module.css
@@ -1,9 +1,6 @@
 .wrapper {
   display: flex;
   flex-direction: column;
+  align-items: flex-start;
   gap: 16px;
-}
-
-.guidelinesButton {
-  align-self: flex-start;
 }
