+++ src/pages/Guideline/Guideline.tsx
@@ -1,11 +1,10 @@
-import { ChevronLeftIcon } from "@radix-ui/react-icons";
-import { useLocation, useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { BookmarkToast } from "../../components/BookmarkToast/BookmarkToast";
-import { Button } from "../../components/Button/Button";
 import { GuidelineDetails } from "../../components/GuidelineDetails/GuidelineDetails";
 import { getCompactId } from "../../utils/guideline";
 import { useGuidelines } from "../../utils/useGuidelines";
 import styles from "./Guideline.module.css";
+import { GuidelinesButton } from "./GuidelinesButton/GuidelinesButton";
 
 export const Guideline = () => {
   const { id } = useParams();
@@ -24,26 +23,5 @@ export const Guideline = () => {
       </div>
       <BookmarkToast />
     </>
-  );
-};
-
-const GuidelinesButton = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  return (
-    <Button
-      theme="ghost"
-      className={styles.guidelinesButton}
-      onClick={() => {
-        if (location.key === "default") {
-          navigate("/");
-        } else {
-          navigate(-1);
-        }
-      }}
-    >
-      <ChevronLeftIcon /> Guidelines
-    </Button>
   );
 };
